GUI: rename all header files from .h to .hpp
